poc/pyproject.toml
[project]
name = "diction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.1",
    "openai>=1.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "websockets>=13.0",
]
